Imperial Command Hat should now be wearable by both Humans and Zabraks
Currently on the Roadmap under GU 16.45 (April 2010 Bug Bash).
I've added the support for the other version of the command hat that is available from the recruiter. Unsure if this was added for the static version on live, but without it even human females wouldn't be able to equip that version (its quite old).