Junk Dealer Lucky Proc
This fixes a credit exploit. If you have an item which sells for a large amount to the Junk Dealer, you can spam sell/buy-back. The Lucky-proc sell price is not taken in to account when buying the item back. If you sell an item for 50mil, and proc Lucky, it sells for 60mil - a profit of 10mil. Buying it back only costs you 50mil. Rinse and repeat.