Updates NPC vendor mobiles to use vendor inventory (not player housing vendors)
This updates unique NPC vendors found in the world to use vendor_inventory's instead of the default creature_inventory. This follows SOE's use for the Aurilia vendors, allowing for a greater item capacity (unlimited). Previously they utilized the creature_inventory which had a capacity of 75 but in practice was closer to 37 items. Only mobiles that were exclusively used as world vendors were updated. Some vendors such as Hoth, Loveday and Meatlumps it may be advised to give them unique mobiles in order to update them to this. This is purely serverside and does not need an update to the client.