GMF Spooky Hand Adjustments
Adjusts Ponda Baba's arm in the master item list to reflect how it was on live, this removes the autostack and 1 time use it previously had. Implemented script side for the spooky hand buff so it correctly gives bonus tokens as it did on live (+1 token per successful horrified scare and +2 tokens per successful terrified scare).